Trying the web api

+++ src/Server/Entities/ClientApp.cs
@@ -5,7 +5,7 @@ namespace Nova.Identity.Entities
         public short Id { get; init; }
         public string Name { get; init; } = "";
         public string? LookupKey { get; init; }
-        public int? BoundaryId { get; init; }
+        public short? BoundaryId { get; init; }
 
         public Boundary? Boundary { get; init; }
     }
